+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Polyverse Boost API on Node.js
 
 ### Enhancements
 - Renamed API from get_file_from_uri to user_project_file (GET)
+- Renamed API from /user_project/ to /user_project/:org/:project
+- Renamed API from /user_project_goals/ to /user_project/:org/:project/goals
+- Project Data and Project Goals are persisted
 
 ### Bug Fixes
 - N/A